// designer/client/reducers/listActions.tsx
export enum ListActions {
  EDIT_TITLE = "EDIT_TITLE",

  SET_SELECTED_LIST = "SET_SELECTED_LIST",
  ADD_NEW_LIST = "ADD_NEW_LIST",

  ADD_LIST_ITEM = "ADD_LIST_ITEM",
  EDIT_LIST_VALUE_TYPE = "EDIT_LIST_VALUE_TYPE",
  EDIT_LIST_ITEM = "EDIT_LIST_ITEM",
  ADD_SUBCOMPONENT = "ADD_SUBCOMPONENT",
  EDIT_SUBCOMPONENT = "EDIT_SUBCOMPONENT",

  EDIT_LIST_ITEM_TEXT = "EDIT_LIST_ITEM_TEXT",
  EDIT_LIST_ITEM_DESCRIPTION = "EDIT_LIST_ITEM_DESCRIPTION",
  EDIT_LIST_ITEM_VALUE = "EDIT_LIST_ITEM_VALUE",
  EDIT_LIST_ITEM_CONDITION = "EDIT_LIST_ITEM_CONDITION",
  SUBMIT_LIST_ITEM = "SUBMIT_LIST_ITEM",
  SUBMIT_SUBCOMPONENT = "SUBMIT_SUBCOMPONENT",

  SUBMIT = "SUBMIT",
}